NXT-574 Create utils folder for all common functions
