+++ web/src/pages/[...slug].page.tsx
@@ -46,6 +46,7 @@ import {
 } from "@/utils/storyblok";
 
 import type { GetServerSidePropsContext } from "next";
+import { fetchParentAndSiblingLinks, filterFunctionForTreeStructure, reUseFetchingLogic } from "@/components/Storyblok/StoryblokSectionNav/utils/Utils";
 
 type Link = {
 	childLinks?: Link[];
@@ -166,94 +167,6 @@ export default function SlugCatchAll(
 		</>
 	);
 }
-const fetchLinksFromStoryblok = async (
-	queryParams: Record<string, string>
-): Promise<Link[]> => {
-	const queryString = new URLSearchParams(queryParams).toString();
-	const res = await fetch(
-		`https://api.storyblok.com/v2/cdn/links?${queryString}`
-	);
-	const data = await res.json();
-	const dataArray: Link[] = Object.values(data.links);
-	return dataArray;
-};
-const fetchParentAndSiblingLinks = async (
-	token: string,
-	parentID: string,
-	slug: string
-) => {
-	// Fetch sibling links first
-	const siblingsLinksArray = await fetchLinksFromStoryblok({
-		version: "published",
-		token,
-		with_parent: parentID || "",
-		per_page: "1000",
-	});
-	const parentAndSiblingLinksArray = await reUseFetchingLogic(
-		token,
-		slug,
-		siblingsLinksArray
-	);
-	return { siblingsLinksArray, parentAndSiblingLinksArray };
-};
-const filterFunctionForTreeStructure = (
-	siblingsLinksArray: Link[],
-	parent: Link
-) => {
-	const children = siblingsLinksArray.filter((childLink) => {
-		const isChild =
-			childLink.parent_id === parent.id && !childLink.is_startpage;
-
-		return isChild;
-	});
-	return children;
-};
-const reUseFetchingLogic = async (
-	token: string,
-	slug: string,
-	siblingsLinksArray: Link[],
-	secondIteration?: boolean
-) => {
-	const startsWithLinksArray = await fetchLinksFromStoryblok({
-		version: "published",
-		token,
-		starts_with: slug,
-		per_page: "1000",
-	});
-
-	const currentFolderLink: Link | undefined = startsWithLinksArray.find(
-		(item: Link) => item.is_folder && item.slug === slug
-	);
-
-	let parentAndSiblingLinksArray: Link[] = [];
-
-	if (currentFolderLink && currentFolderLink.parent_id) {
-		parentAndSiblingLinksArray = await fetchLinksFromStoryblok({
-			version: "published",
-			token,
-			with_parent: currentFolderLink.parent_id,
-			per_page: "1000",
-		});
-		if (secondIteration) {
-			parentAndSiblingLinksArray.map((parentelse) => {
-				const children = filterFunctionForTreeStructure(
-					siblingsLinksArray,
-					parentelse
-				);
-				parentelse.childLinks = children;
-				if (children.length > 0) {
-					parentelse.childLinks = children;
-				} else {
-					parentelse.childLinks = [];
-				}
-			});
-		}
-	} else {
-		parentAndSiblingLinksArray = siblingsLinksArray;
-	}
-	return parentAndSiblingLinksArray;
-};
-
 export async function getServerSideProps(context: GetServerSidePropsContext) {
 	// Bail out early unless this route is enabled for this environment
 	if (publicRuntimeConfig.storyblok.enableRootCatchAll.toString() !== "true") {
